+++ src/components/BsModal/index.js
@@ -1,7 +1,0 @@
-import BsModal from "./BsModal";
-import BsLightbox from "./BsLightbox";
-
-export default Vue => {
-    Vue.component(BsModal.name, BsModal);
-    Vue.component(BsLightbox.name, BsLightbox);
-};